--- a/aot/src/main/java/com/dylibso/chicory/aot/AotMachine.java
+++ b/aot/src/main/java/com/dylibso/chicory/aot/AotMachine.java
@@ -73,6 +73,7 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.commons.InstructionAdapter;
 import org.objectweb.asm.util.CheckClassAdapter;
+import org.objectweb.asm.util.TraceClassVisitor;
 
 /**
  * Simple Machine implementation that AOT compiles function bodies and runtime-links them
@@ -511,8 +512,11 @@ public final class AotMachine implements Machine {
         } catch (VerifyError e) {
             // run ASM verifier to help with debugging
             try {
+                PrintWriter pw = new PrintWriter(System.out, true, UTF_8);
                 ClassReader reader = new ClassReader(classBytes);
-                CheckClassAdapter.verify(reader, true, new PrintWriter(System.out, false, UTF_8));
+                CheckClassAdapter.verify(reader, true, pw);
+
+                reader.accept(new TraceClassVisitor(pw), 0);
             } catch (NoClassDefFoundError ignored) {
                 // the ASM verifier is an optional dependency
             } catch (Throwable t) {
@@ -589,9 +593,10 @@ public final class AotMachine implements Machine {
     }
 
     private static void emitBoxArguments(MethodVisitor asm, List<ValueType> types) {
+        int slotCount = types.stream().mapToInt(AotUtil::slotCount).sum();
         int slot = 0;
         // box the arguments into long[]
-        asm.visitLdcInsn(types.size());
+        asm.visitLdcInsn(slotCount);
         asm.visitIntInsn(Opcodes.NEWARRAY, Opcodes.T_LONG); // long
         for (int i = 0; i < types.size(); i++) {
             asm.visitInsn(Opcodes.DUP);