… README.md
… aot-tests/pom.xml
… aot-tests/src/test/java/com/dylibso/chicory/testing/Spectest.java
… aot/pom.xml
… aot/src/main/java/com/dylibso/chicory/aot/AotUtil.java
… aot/src/main/java/com/dylibso/chicory/aot/HostFunctionInvoker.java
… aot/src/main/java/com/dylibso/chicory/aot/LongArrayWrapper.java
… aot/src/main/java/com/dylibso/chicory/aot/ValueConversions.java
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyBrTable.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyBranching.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyFloat.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyHelloWasi.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyI32.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyI32Renamed.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyIterFact.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyKitchenSink.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyMemory.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyStart.approved.txt
… aot/src/test/resources/com/dylibso/chicory/approvals/ApprovalTest.verifyTrap.approved.txt
… cli/src/main/java/com/dylibso/chicory/cli/Cli.java
… function-processor/src/main/java/com/dylibso/chicory/function/processor/FunctionProcessor.java
… function-processor/src/test/resources/BasicMathGenerated.java
… function-processor/src/test/resources/NestedGenerated.java
… function-processor/src/test/resources/NoPackageGenerated.java
… function-processor/src/test/resources/SimpleGenerated.java
… jmh/src/main/java/com/dylibso/chicory/bench/BenchmarkFactorialExecution.java
… pom.xml
… runtime-tests/src/test/java/com/dylibso/chicory/testing/Spectest.java
… runtime/src/main/java/com/dylibso/chicory/runtime/ConstantEvaluators.java
… runtime/src/main/java/com/dylibso/chicory/runtime/ExportFunction.java