Supersede #447 this is a complete implementation, already passing all the tests up to the interpreter, can be tested with something like:

```
 mvn clean install -pl runtime-tests -am
```

I started, but got a bit confused about the changes necessary in "aot", @electrum or @danielperano I would love some help in finalizing this PR :pray: and learning a bit more how `asm` works ...

This is a prerequisite to avoid arrays construction [here](https://github.com/dylibso/chicory/pull/528/files#diff-f9f476fefd928e6770830c090018da2b95a054834c70bf472460d568fd1a9d2bR184), the plan is to open up the underlying array so that we can perform operations just using the offset :crossed_fingers: