@andreaTP 2024-09-19
update: I benchmarked the interpreter and got good improvements on this branch, I think it's in the right direction!
Regarding the AOT, the changes seem to be a bit more invasive than expected, it would be great to have some help.