@andreaTP 2024-09-25
I did a little benchmarking on my super noisy machine(take the results with a lot of salt):

```
Chicory 0.0.12:

With GraalJIT:
Benchmark                                      Mode  Cnt      Score        Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3   1282.401 ±    51.518  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  63086.906 ± 18526.425  ns/op

Without GraalJIT:
Benchmark                                      Mode  Cnt      Score        Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3   1222.858 ±    326.915  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  87630.869 ± 138021.856  ns/op

---

Current main:

With GraalJIT:
Benchmark                                      Mode  Cnt      Score       Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3   1237.301 ±   423.694  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  71274.055 ± 11410.525  ns/op

Without GraalJIT:
Benchmark                                      Mode  Cnt       Score       Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3    1349.806 ±   945.091  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  101900.141 ± 38884.734  ns/op

---

Primitive longs (this branch):

With GraalJIT:
Benchmark                                      Mode  Cnt      Score       Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3   1153.859 ±   984.977  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  41791.372 ± 43360.065  ns/op

Without GraalJIT:
Benchmark                                      Mode  Cnt      Score       Error  Units
ChicoryAotTest.chicoryAotTest                  avgt    3    899.077 ±   483.678  ns/op
ChicoryInterpreterTest.chicoryInterpreterTest  avgt    3  49629.772 ± 88844.628  ns/op

```